// src/generated/accounts/loan.ts
export const LOAN_DISCRIMINATOR = new Uint8Array([
  20, 195, 70, 117, 165, 227, 182, 1,
]);

export function getLoanDiscriminatorBytes() {
  return fixEncoderSize(getBytesEncoder(), 8).encode(LOAN_DISCRIMINATOR);
}

export type Loan = {
  discriminator: ReadonlyUint8Array;
  version: number;
  /** amount in tokens, decimals included */
  principalLamports: bigint;
  orderBook: Address;
  /** Token mint for what the loan is in (spl address) */
  valueTokenMint: Address;
  escrowBumpSeed: number;
  /** stores start and duration */
  loanState: LoanState;
};